Edited code to fit an ArtCollection class.
